Restructured the frontend codebase
