+++ client_app/src/Components/shelter/NavBarShelterDashboard.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 import { Navbar, Nav } from "react-bootstrap";
-import "./NavBar.css";
+import "../../styles/NavBar.css";
 import { useLocation } from "react-router-dom";
 
-function CustomNavBar({ auth }) {
+function CustomNavBarShelter({ auth }) {
   const location = useLocation();
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   useEffect(() => {
@@ -18,23 +18,27 @@ function CustomNavBar({ auth }) {
   return (
     <>
       <Navbar collapseOnSelect expand="lg" sticky="top" data-bs-theme="dark">
-        <Navbar.Brand href="/">Volunteer Dashboard</Navbar.Brand>
+        <Navbar.Brand href="/shelter-dashboard">Shelter Dashboard</Navbar.Brand>
         <Navbar.Toggle aria-controls="responsive-navbar-nav" />
         <Navbar.Collapse id="responsive-navbar-nav">
           <Nav className="justify-content-end" style={{ width: "100%" }}>
             {isAuthenticated || auth ? (
               <>
-                <Nav.Link href="/" active={location.pathname === "/dashboard"}>
-                  Your Dashboard
+                <Nav.Link
+                  href="/shelter-dashboard"
+                  active={location.pathname === "/shelter-dashboard"}>
+                  Shelter Dashboard
                 </Nav.Link>
-                <Nav.Link href="/shelters" active={location.pathname === "/shelters"}>
-                  Sign Up For Shifts
+                <Nav.Link href="/shift-details" active={location.pathname === "/shift-details"}>
+                  Shift Details
                 </Nav.Link>
-                <Nav.Link href="/upcoming-shifts" active={location.pathname === "/upcoming-shifts"}>
-                  Your Upcoming Shifts
+                <Nav.Link
+                  href="/request-for-help"
+                  active={location.pathname === "/request-for-help"}>
+                  Request For Help
                 </Nav.Link>
-                <Nav.Link href="/past-shifts" active={location.pathname === "/past-shifts"}>
-                  Your Previous Shifts
+                <Nav.Link href="/cancel-shifts" active={location.pathname === "/cancel-shifts"}>
+                  Cancel Shifts
                 </Nav.Link>
                 <Nav.Link href="/logout">Sign Out</Nav.Link>
               </>
@@ -51,4 +55,4 @@ function CustomNavBar({ auth }) {
   );
 }
 
-export default CustomNavBar;
+export default CustomNavBarShelter;
