+++ client_app/src/ShelterDashboard.js
@@ -1,7 +1,7 @@
 import React from "react";
-import OpenRequest from "./OpenRequests";
-import PastVolunteersContainer from "./PastVolunteersContainer";
-import ShiftContainer from "./ShiftContainer";
+import OpenRequest from "./components/shelter/OpenRequests";
+import PastVolunteersContainer from "./components/shelter/PastVolunteersContainer";
+import ShiftContainer from "./components/shelter/ShiftContainer";
 import "./index.css";
 
 function ShelterDashboard() {
